# Conflicts:
#	protocol/http/service/player.go

+++ entity/benchmark_test.go
@@ -31,7 +31,7 @@ func BenchmarkTestCostEquals(b *testing.B) {
 
 func BenchmarkTestPlayerChainNext(b *testing.B) {
 	pc := newPlayerChain()
-	for i := uint(0); i < 100; i++ {
+	for i := uint64(0); i < 100; i++ {
 		pc.add(i)
 	}
 
@@ -43,14 +43,14 @@ func BenchmarkTestPlayerChainNext(b *testing.B) {
 
 func BenchmarkTestPlayerChainNextWithComplete(b *testing.B) {
 	pc := newPlayerChain()
-	for i := uint(0); i < 11451419; i++ {
+	for i := uint64(0); i < 11451419; i++ {
 		pc.add(i)
 	}
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		exist, _ := pc.next()
-		pc.complete(uint(i))
+		pc.complete(uint64(i))
 		if !exist {
 			break
 		}
@@ -60,9 +60,9 @@ func BenchmarkTestPlayerChainNextWithComplete(b *testing.B) {
 func BenchmarkTestEventMapPreview(b *testing.B) {
 	m := NewEventMap()
 	for i := 0; i < 128; i++ {
-		m.AddEvent(newTestEvent(uint(i), enum.AfterAttack, true, func(ctx *context.CallbackContext) { ctx.SwitchCharacter(114514) }))
+		m.AddEvent(newTestEvent(uint64(i), enum.AfterAttack, true, func(ctx *context.CallbackContext) { ctx.SwitchCharacter(114514) }))
 	}
-	m.AddEvent(newTestEvent(uint(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) }))
+	m.AddEvent(newTestEvent(uint64(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) }))
 	ctx := context.NewCallbackContext()
 	b.ResetTimer()
 
@@ -74,9 +74,9 @@ func BenchmarkTestEventMapPreview(b *testing.B) {
 func BenchmarkTestEventMapExecute(b *testing.B) {
 	m := NewEventMap()
 	for i := 0; i < 128; i++ {
-		m.AddEvent(newTestEvent(uint(i), enum.AfterAttack, true, func(ctx *context.CallbackContext) { ctx.SwitchCharacter(114514) }))
+		m.AddEvent(newTestEvent(uint64(i), enum.AfterAttack, true, func(ctx *context.CallbackContext) { ctx.SwitchCharacter(114514) }))
 	}
-	m.AddEvent(newTestEvent(uint(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) }))
+	m.AddEvent(newTestEvent(uint64(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) }))
 	ctx := context.NewCallbackContext()
 	b.ResetTimer()
 
@@ -90,13 +90,13 @@ func BenchmarkTestEventMapAdd(b *testing.B) {
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
-		m.AddEvent(newTestEvent(uint(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) }))
+		m.AddEvent(newTestEvent(uint64(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) }))
 	}
 }
 
 func BenchmarkTestEventMapRemove(b *testing.B) {
 	m := NewEventMap()
-	event := newTestEvent(uint(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) })
+	event := newTestEvent(uint64(1), enum.AfterSwitch, true, func(ctx *context.CallbackContext) { ctx.ChangeOperated(false) })
 	m.AddEvent(event)
 	b.ResetTimer()
